feat!(core): add `metadata` mod
