+++ crates/llama-core/src/lib.rs
@@ -11,6 +11,7 @@ pub mod embeddings;
 pub mod error;
 pub mod graph;
 pub mod images;
+pub mod metadata;
 pub mod models;
 pub mod rag;
 #[cfg(feature = "search")]
@@ -19,10 +20,9 @@ pub mod utils;
 
 pub use error::LlamaCoreError;
 pub use graph::{EngineType, Graph, GraphBuilder};
+pub use metadata::{ggml::GgmlMetadata, whisper::WhisperMetadata, BaseMetadata};
 
-use chat_prompts::PromptTemplateType;
 use once_cell::sync::OnceCell;
-use serde::{Deserialize, Serialize};
 use std::{
     collections::HashMap,
     path::Path,
@@ -53,462 +53,6 @@ pub(crate) static PIPER_GRAPH: OnceCell<Mutex<Graph<GgmlMetadata>>> = OnceCell::
 pub(crate) const MAX_BUFFER_SIZE: usize = 2usize.pow(14) * 15 + 128;
 pub(crate) const OUTPUT_TENSOR: usize = 0;
 const PLUGIN_VERSION: usize = 1;
-
-pub trait BaseMetadata {
-    fn model_name(&self) -> &str;
-    fn model_alias(&self) -> &str;
-    fn prompt_template(&self) -> PromptTemplateType;
-}
-
-/// Metadata for chat and embeddings models
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct GgmlMetadata {
-    // this field not defined for the beckend plugin
-    #[serde(skip_serializing)]
-    pub model_name: String,
-    // this field not defined for the beckend plugin
-    #[serde(skip_serializing)]
-    pub model_alias: String,
-    // this field not defined for the beckend plugin
-    #[serde(skip_serializing)]
-    pub log_prompts: bool,
-    // this field not defined for the beckend plugin
-    #[serde(skip_serializing)]
-    pub prompt_template: PromptTemplateType,
-
-    // * Plugin parameters (used by this plugin):
-    #[serde(rename = "enable-log")]
-    pub log_enable: bool,
-    #[serde(rename = "enable-debug-log")]
-    pub debug_log: bool,
-    // #[serde(rename = "stream-stdout")]
-    // pub stream_stdout: bool,
-    #[serde(rename = "embedding")]
-    pub embeddings: bool,
-    #[serde(rename = "n-predict")]
-    pub n_predict: u64,
-    #[serde(skip_serializing_if = "Option::is_none", rename = "reverse-prompt")]
-    pub reverse_prompt: Option<String>,
-    /// path to the multimodal projector file for llava
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub mmproj: Option<String>,
-    /// Path to the image file for llava
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub image: Option<String>,
-
-    // * Model parameters (need to reload the model if updated):
-    #[serde(rename = "n-gpu-layers")]
-    pub n_gpu_layers: u64,
-    /// The main GPU to use. Defaults to None.
-    #[serde(rename = "main-gpu")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub main_gpu: Option<u64>,
-    /// How split tensors should be distributed accross GPUs. If None the model is not split; otherwise, a comma-separated list of non-negative values, e.g., "3,2" presents 60% of the data to GPU 0 and 40% to GPU 1.
-    #[serde(rename = "tensor-split")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tensor_split: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none", rename = "use-mmap")]
-    pub use_mmap: Option<bool>,
-    // * Context parameters (used by the llama context):
-    #[serde(rename = "ctx-size")]
-    pub ctx_size: u64,
-    #[serde(rename = "batch-size")]
-    pub batch_size: u64,
-    #[serde(rename = "threads")]
-    pub threads: u64,
-
-    // * Sampling parameters (used by the llama sampling context).
-    #[serde(rename = "temp")]
-    pub temperature: f64,
-    #[serde(rename = "top-p")]
-    pub top_p: f64,
-    #[serde(rename = "repeat-penalty")]
-    pub repeat_penalty: f64,
-    #[serde(rename = "presence-penalty")]
-    pub presence_penalty: f64,
-    #[serde(rename = "frequency-penalty")]
-    pub frequency_penalty: f64,
-
-    // * grammar parameters
-    /// BNF-like grammar to constrain generations (see samples in grammars/ dir). Defaults to empty string.
-    pub grammar: String,
-    /// JSON schema to constrain generations (<https://json-schema.org/>), e.g. `{}` for any JSON object. For schemas w/ external $refs, use --grammar + example/json_schema_to_grammar.py instead.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub json_schema: Option<String>,
-}
-impl Default for GgmlMetadata {
-    fn default() -> Self {
-        Self {
-            model_name: String::new(),
-            model_alias: String::new(),
-            log_prompts: false,
-            debug_log: false,
-            prompt_template: PromptTemplateType::Llama2Chat,
-            log_enable: false,
-            embeddings: false,
-            n_predict: 1024,
-            reverse_prompt: None,
-            mmproj: None,
-            image: None,
-            n_gpu_layers: 100,
-            main_gpu: None,
-            tensor_split: None,
-            use_mmap: Some(true),
-            ctx_size: 512,
-            batch_size: 512,
-            threads: 2,
-            temperature: 1.0,
-            top_p: 1.0,
-            repeat_penalty: 1.1,
-            presence_penalty: 0.0,
-            frequency_penalty: 0.0,
-            grammar: String::new(),
-            json_schema: None,
-        }
-    }
-}
-impl BaseMetadata for GgmlMetadata {
-    fn model_name(&self) -> &str {
-        &self.model_name
-    }
-
-    fn model_alias(&self) -> &str {
-        &self.model_alias
-    }
-
-    fn prompt_template(&self) -> PromptTemplateType {
-        self.prompt_template
-    }
-}
-
-/// Builder for creating a ggml metadata
-#[derive(Debug)]
-pub struct GgmlMetadataBuilder {
-    metadata: GgmlMetadata,
-}
-impl GgmlMetadataBuilder {
-    pub fn new<S: Into<String>>(model_name: S, model_alias: S, pt: PromptTemplateType) -> Self {
-        let metadata = GgmlMetadata {
-            model_name: model_name.into(),
-            model_alias: model_alias.into(),
-            prompt_template: pt,
-            ..Default::default()
-        };
-
-        Self { metadata }
-    }
-
-    pub fn with_prompt_template(mut self, template: PromptTemplateType) -> Self {
-        self.metadata.prompt_template = template;
-        self
-    }
-
-    pub fn enable_plugin_log(mut self, enable: bool) -> Self {
-        self.metadata.log_enable = enable;
-        self
-    }
-
-    pub fn enable_debug_log(mut self, enable: bool) -> Self {
-        self.metadata.debug_log = enable;
-        self
-    }
-
-    pub fn enable_prompts_log(mut self, enable: bool) -> Self {
-        self.metadata.log_prompts = enable;
-        self
-    }
-
-    pub fn enable_embeddings(mut self, enable: bool) -> Self {
-        self.metadata.embeddings = enable;
-        self
-    }
-
-    pub fn with_n_predict(mut self, n: u64) -> Self {
-        self.metadata.n_predict = n;
-        self
-    }
-
-    pub fn with_main_gpu(mut self, gpu: Option<u64>) -> Self {
-        self.metadata.main_gpu = gpu;
-        self
-    }
-
-    pub fn with_tensor_split(mut self, split: Option<String>) -> Self {
-        self.metadata.tensor_split = split;
-        self
-    }
-
-    pub fn with_threads(mut self, threads: u64) -> Self {
-        self.metadata.threads = threads;
-        self
-    }
-
-    pub fn with_reverse_prompt(mut self, prompt: Option<String>) -> Self {
-        self.metadata.reverse_prompt = prompt;
-        self
-    }
-
-    pub fn with_mmproj(mut self, path: Option<String>) -> Self {
-        self.metadata.mmproj = path;
-        self
-    }
-
-    pub fn with_image(mut self, path: impl Into<String>) -> Self {
-        self.metadata.image = Some(path.into());
-        self
-    }
-
-    pub fn with_n_gpu_layers(mut self, n: u64) -> Self {
-        self.metadata.n_gpu_layers = n;
-        self
-    }
-
-    pub fn disable_mmap(mut self, disable: Option<bool>) -> Self {
-        self.metadata.use_mmap = disable.map(|v| !v);
-        self
-    }
-
-    pub fn with_ctx_size(mut self, size: u64) -> Self {
-        self.metadata.ctx_size = size;
-        self
-    }
-
-    pub fn with_batch_size(mut self, size: u64) -> Self {
-        self.metadata.batch_size = size;
-        self
-    }
-
-    pub fn with_temperature(mut self, temp: f64) -> Self {
-        self.metadata.temperature = temp;
-        self
-    }
-
-    pub fn with_top_p(mut self, top_p: f64) -> Self {
-        self.metadata.top_p = top_p;
-        self
-    }
-
-    pub fn with_repeat_penalty(mut self, penalty: f64) -> Self {
-        self.metadata.repeat_penalty = penalty;
-        self
-    }
-
-    pub fn with_presence_penalty(mut self, penalty: f64) -> Self {
-        self.metadata.presence_penalty = penalty;
-        self
-    }
-
-    pub fn with_frequency_penalty(mut self, penalty: f64) -> Self {
-        self.metadata.frequency_penalty = penalty;
-        self
-    }
-
-    pub fn with_grammar(mut self, grammar: impl Into<String>) -> Self {
-        self.metadata.grammar = grammar.into();
-        self
-    }
-
-    pub fn with_json_schema(mut self, schema: Option<String>) -> Self {
-        self.metadata.json_schema = schema;
-        self
-    }
-
-    pub fn build(self) -> GgmlMetadata {
-        self.metadata
-    }
-}
-
-/// Metadata for whisper model
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct WhisperMetadata {
-    // this field not defined for the beckend plugin
-    #[serde(skip_serializing)]
-    pub model_name: String,
-    // this field not defined for the beckend plugin
-    #[serde(skip_serializing)]
-    pub model_alias: String,
-
-    /// Enable debug mode. Defaults to false.
-    pub debug_mode: bool,
-    /// Number of threads to use during computation. Defaults to 4.
-    pub threads: u64,
-    /// Translate from source language to english. Defaults to false.
-    pub translate: bool,
-    /// The language of the input audio. `auto` for auto-detection. Defaults to `en`.
-    ///
-    /// Supplying the input language in [ISO-639-1](https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes) format will improve accuracy and latency.
-    pub language: String,
-    /// Number of processors to use during computation. Defaults to 1.
-    pub processors: u32,
-    /// Time offset in milliseconds. Defaults to 0.
-    pub offset_t: u32,
-    /// Duration of audio to process in milliseconds. Defaults to 0.
-    pub duration: u32,
-    /// Maximum number of text context tokens to store. Defaults to -1.
-    pub max_context: i32,
-    /// Maximum segment length in characters. Defaults to 0.
-    pub max_len: u32,
-    /// Split on word rather than on token. Defaults to false.
-    pub split_on_word: bool,
-    /// Output result in a text file. Defaults to false.
-    pub output_txt: bool,
-    /// Output result in a vtt file. Defaults to false.
-    pub output_vtt: bool,
-    /// Output result in a srt file. Defaults to false.
-    pub output_srt: bool,
-    /// Output result in a lrc file. Defaults to false.
-    pub output_lrc: bool,
-    /// Output result in a CSV file. Defaults to false.
-    pub output_csv: bool,
-    /// Output result in a JSON file. Defaults to false.
-    pub output_json: bool,
-    /// The sampling temperature, between 0 and 1. Defaults to 0.00.
-    pub temperature: f64,
-}
-impl Default for WhisperMetadata {
-    fn default() -> Self {
-        Self {
-            model_name: String::new(),
-            model_alias: String::new(),
-            debug_mode: false,
-            threads: 4,
-            translate: false,
-            language: "en".to_string(),
-            processors: 1,
-            offset_t: 0,
-            duration: 0,
-            max_context: -1,
-            max_len: 0,
-            split_on_word: false,
-            output_txt: false,
-            output_vtt: false,
-            output_srt: false,
-            output_lrc: false,
-            output_csv: false,
-            output_json: false,
-            temperature: 0.0,
-        }
-    }
-}
-impl BaseMetadata for WhisperMetadata {
-    fn model_name(&self) -> &str {
-        &self.model_name
-    }
-
-    fn model_alias(&self) -> &str {
-        &self.model_alias
-    }
-
-    fn prompt_template(&self) -> PromptTemplateType {
-        PromptTemplateType::Null
-    }
-}
-
-/// Builder for creating an audio metadata
-#[derive(Debug)]
-pub struct WhisperMetadataBuilder {
-    metadata: WhisperMetadata,
-}
-impl WhisperMetadataBuilder {
-    pub fn new<S: Into<String>>(model_name: S, model_alias: S) -> Self {
-        let metadata = WhisperMetadata {
-            model_name: model_name.into(),
-            model_alias: model_alias.into(),
-            ..Default::default()
-        };
-
-        Self { metadata }
-    }
-
-    pub fn enable_debug(mut self, enable: bool) -> Self {
-        self.metadata.debug_mode = enable;
-        self
-    }
-
-    pub fn with_threads(mut self, threads: u64) -> Self {
-        self.metadata.threads = threads;
-        self
-    }
-
-    pub fn enable_translate(mut self, enable: bool) -> Self {
-        self.metadata.translate = enable;
-        self
-    }
-
-    pub fn with_language(mut self, language: String) -> Self {
-        self.metadata.language = language;
-        self
-    }
-
-    pub fn with_processors(mut self, processors: u32) -> Self {
-        self.metadata.processors = processors;
-        self
-    }
-
-    pub fn with_offset_t(mut self, offset_t: u32) -> Self {
-        self.metadata.offset_t = offset_t;
-        self
-    }
-
-    pub fn with_duration(mut self, duration: u32) -> Self {
-        self.metadata.duration = duration;
-        self
-    }
-
-    pub fn with_max_context(mut self, max_context: i32) -> Self {
-        self.metadata.max_context = max_context;
-        self
-    }
-
-    pub fn with_max_len(mut self, max_len: u32) -> Self {
-        self.metadata.max_len = max_len;
-        self
-    }
-
-    pub fn split_on_word(mut self, split_on_word: bool) -> Self {
-        self.metadata.split_on_word = split_on_word;
-        self
-    }
-
-    pub fn output_txt(mut self, output_txt: bool) -> Self {
-        self.metadata.output_txt = output_txt;
-        self
-    }
-
-    pub fn output_vtt(mut self, output_vtt: bool) -> Self {
-        self.metadata.output_vtt = output_vtt;
-        self
-    }
-
-    pub fn output_srt(mut self, output_srt: bool) -> Self {
-        self.metadata.output_srt = output_srt;
-        self
-    }
-
-    pub fn output_lrc(mut self, output_lrc: bool) -> Self {
-        self.metadata.output_lrc = output_lrc;
-        self
-    }
-
-    pub fn output_csv(mut self, output_csv: bool) -> Self {
-        self.metadata.output_csv = output_csv;
-        self
-    }
-
-    pub fn output_json(mut self, output_json: bool) -> Self {
-        self.metadata.output_json = output_json;
-        self
-    }
-
-    pub fn with_temperature(mut self, temperature: f64) -> Self {
-        self.metadata.temperature = temperature;
-        self
-    }
-
-    pub fn build(self) -> WhisperMetadata {
-        self.metadata
-    }
-}
 
 /// Initialize the core context
 pub fn init_core_context(
