+++ crates/llama-core/src/metadata/mod.rs
@@ -1,0 +1,7 @@
+pub mod ggml;
+pub mod whisper;
+
+pub trait BaseMetadata {
+    fn model_name(&self) -> &str;
+    fn model_alias(&self) -> &str;
+}
